Machine Learning/Code/requirements.txt
huggingface_hub==0.36.0
matplotlib==3.10.8
numpy==2.4.0
ollama==0.6.1
opencv_python==4.8.1.78
opencv_python_headless==4.10.0.84
pandas==2.3.3
Pillow==12.0.0
sam3==0.1.2
scikit_learn==1.8.0
seaborn==0.13.2
torch==2.5.1+cu121
tqdm==4.67.1
ultralytics==8.3.239
